requirements.txt
streamlit
opencv-python-headless>=4.5.0
numpy>=1.21.0
streamlit>=1.24.0
pillow>=9.0.0
scikit-learn>=1.0.0
face_recognition>=1.3.0
pytesseract>=0.3.8
matplotlib>=3.4.0
seaborn>=0.11.0
requests>=2.26.0


